move accessors code to separate file
